third_party/WebKit/LayoutTests/inspector/sources/debugger-breakpoints/event-listener-breakpoints.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>

function testElementEventListener()
{
    return 0;
}

function addListenerAndClick()
{
    var element = document.getElementById("test");
    element.addEventListener("click", testElementEventListener, true);
    element.click();
}

function addListenerAndAuxclick()
{
    var element = document.getElementById("test");
    element.addEventListener("auxclick", testElementEventListener, true);
    element.dispatchEvent(new MouseEvent("auxclick", {button: 1}));
}

function addListenerAndPointerDown()
{
    var element = document.getElementById("test");
    element.addEventListener("pointerdown", testElementEventListener, true);
    element.dispatchEvent(new PointerEvent("pointerdown"));
}

function timerFired()
{
    return 0;
}

function addLoadListeners()
{
    var xhr = new XMLHttpRequest();
    xhr.onload = loadCallback;
    xhr.onerror = loadCallback;
    xhr.open("GET", "http://localhost/", true);

    var img = new Image();
    img.onload = sendXHR;
    img.onerror = sendXHR;
    img.src = "foo/bar/dummy";

    function sendXHR()
    {
        xhr.send();
    }
}

function loadCallback()
{
    return 0;
}

function playVideo()
{
    var video = document.getElementById("video");
    video.addEventListener("play", onVideoPlay, false);
    video.play();
}

function onVideoPlay()
{
    return 0;
}

function test()
{
    var testFunctions = [
        function testClickBreakpoint(next)
        {
            InspectorTest.setEventListenerBreakpoint("listener:click", true);
            InspectorTest.waitUntilPaused(paused);
            InspectorTest.evaluateInPageWithTimeout("addListenerAndClick()");

            function paused(callFrames, reason, breakpointIds, asyncStackTrace, auxData)
            {
                InspectorTest.captureStackTrace(callFrames);
                printEventTargetName(auxData);
                InspectorTest.setEventListenerBreakpoint("listener:click", false);
                InspectorTest.resumeExecution(resumed);
            }

            function resumed()
            {
                InspectorTest.evaluateInPage("addListenerAndClick()", next);
            }
        },

        function testAuxclickBreakpoint(next)
        {
            InspectorTest.setEventListenerBreakpoint("listener:auxclick", true);
            InspectorTest.waitUntilPaused(paused);
            InspectorTest.evaluateInPageWithTimeout("addListenerAndAuxclick()");

            function paused(callFrames, reason, breakpointIds, asyncStackTrace, auxData)
            {
                InspectorTest.captureStackTrace(callFrames);
                printEventTargetName(auxData);
                InspectorTest.setEventListenerBreakpoint("listener:auxclick", false);
                InspectorTest.resumeExecution(resumed);
            }

            function resumed()
            {
                InspectorTest.evaluateInPage("addListenerAndAuxclick()", next);
            }
        },

        function testTimerFiredBreakpoint(next)
        {
            InspectorTest.setEventListenerBreakpoint("instrumentation:setTimeout.callback", true);
            InspectorTest.waitUntilPaused(paused);
            InspectorTest.evaluateInPage("setTimeout(timerFired, 10)");

            function paused(callFrames)
            {
                InspectorTest.captureStackTrace(callFrames);
                InspectorTest.setEventListenerBreakpoint("instrumentation:setTimeout.callback", false);
                InspectorTest.resumeExecution(next);
            }
        },

        function testLoadBreakpointOnXHR(next)
        {
            InspectorTest.setEventListenerBreakpoint("listener:load", true, "xmlhttprequest");
            InspectorTest.setEventListenerBreakpoint("listener:error", true, "xmlhttprequest");
            InspectorTest.waitUntilPaused(paused);
            InspectorTest.evaluateInPageWithTimeout("addLoadListeners()");

            function paused(callFrames, reason, breakpointIds, asyncStackTrace, auxData)
            {
                InspectorTest.captureStackTrace(callFrames);
                printEventTargetName(auxData);
                InspectorTest.setEventListenerBreakpoint("listener:load", false, "xmlhttprequest");
                InspectorTest.setEventListenerBreakpoint("listener:error", false, "xmlhttprequest");
                InspectorTest.resumeExecution(resumed);
            }

            function resumed()
            {
                InspectorTest.evaluateInPage("addLoadListeners()", next);
            }
        },

        function testMediaEventBreakpoint(next)
        {
            InspectorTest.setEventListenerBreakpoint("listener:play", true, "audio");
            InspectorTest.waitUntilPaused(paused);
            InspectorTest.evaluateInPageWithTimeout("playVideo()");

            function paused(callFrames, reason, breakpointIds, asyncStackTrace, auxData)
            {
                InspectorTest.captureStackTrace(callFrames);
                printEventTargetName(auxData);
                InspectorTest.setEventListenerBreakpoint("listener:play", false, "audio");
                InspectorTest.resumeExecution(next);
            }
        }
    ];

    if (window.PointerEvent) {
        testFunctions.push(
            function testPointerEventBreakpoint(next)
            {
                InspectorTest.setEventListenerBreakpoint("listener:pointerdown", true);
                InspectorTest.waitUntilPaused(paused);
                InspectorTest.evaluateInPageWithTimeout("addListenerAndPointerDown()");

                function paused(callFrames, reason, breakpointIds, asyncStackTrace, auxData)
                {
                    InspectorTest.captureStackTrace(callFrames);
                    printEventTargetName(auxData);
                    InspectorTest.setEventListenerBreakpoint("listener:pointerdown", false);
                    InspectorTest.resumeExecution(resumed);
                }

                function resumed()
                {
                    InspectorTest.evaluateInPage("addListenerAndPointerDown()", next);
                }
            }
        );
    }

    InspectorTest.runDebuggerTestSuite(testFunctions);

    function printEventTargetName(auxData)
    {
        var targetName = auxData && auxData.targetName;
        if (targetName)
            InspectorTest.addResult("Event target: " + targetName);
        else
            InspectorTest.addResult("FAIL: No event target name received!");
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests event listener breakpoints.
</p>

<input type=button id="test"></input>
<video id="video" src="../../../media/content/test.ogv"></video>

</body>
</html>
